requirements-cpu.txt
# CPU-only requirements for Neural Operator Foundation Lab

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Data Handling
h5py>=3.9.0
zarr>=2.16.0
dask[complete]>=2023.10.0

# Configuration and Utilities
pyyaml>=6.0
dataclasses-json>=0.6.0
omegaconf>=2.3.0

# Monitoring and Logging
tensorboard>=2.14.0
prometheus-client>=0.18.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7

# Caching
redis>=5.0.0

# Cloud Integration
boto3>=1.29.0
s3fs>=2023.9.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.12.0

# Security
cryptography>=41.0.0